fix: drag & drop branches & loops
