+++ packages/backend/src/app/flows/flow-version/flow-version.service.ts
@@ -2,7 +2,6 @@ import { TSchema, Type } from '@sinclair/typebox'
 import { TypeCompiler } from '@sinclair/typebox/compiler'
 import { PiecePropertyMap, PropertyType } from '@activepieces/pieces-framework'
 import {
-    Action,
     ActionType,
     apId,
     BranchActionSettingsWithValidation,
@@ -18,8 +17,6 @@ import {
     PieceActionSettings,
     PieceTriggerSettings,
     ProjectId,
-    StepLocationRelativeToParent,
-    Trigger,
     TriggerType,
 } from '@activepieces/shared'
 import { QueryDeepPartialEntity } from 'typeorm/query-builder/QueryPartialEntity'
@@ -63,7 +60,7 @@ export const flowVersionService = {
                         type: FlowOperationType.UPDATE_TRIGGER,
                         request: trigger,
                     })
-                    operations.push(...getImportOperations(trigger))
+                    operations.push(...flowHelper.getImportOperations(trigger))
                 }
                 break
             }
@@ -140,106 +137,6 @@ export const flowVersionService = {
         }
         return await flowVersionRepo.save(flowVersion)
     },
-}
-
-function getImportOperations(step: Action | Trigger | undefined): (FlowOperationRequest)[] {
-    const steps: FlowOperationRequest[] = []
-    while (step) {
-        if (step.nextAction) {
-            steps.push({
-                type: FlowOperationType.ADD_ACTION,
-                request: {
-                    parentStep: step.name,
-                    action: keepBaseAction(step.nextAction),
-                },
-            })
-        }
-        if (step.type === ActionType.BRANCH) {
-            if (step.onFailureAction) {
-                steps.push({
-                    type: FlowOperationType.ADD_ACTION,
-                    request: {
-                        parentStep: step.name,
-                        stepLocationRelativeToParent: StepLocationRelativeToParent.INSIDE_FALSE_BRANCH,
-                        action: keepBaseAction(step.onFailureAction),
-                    },
-                })
-                steps.push(...getImportOperations(step.onFailureAction))
-            }
-            if (step.onSuccessAction) {
-                steps.push({
-                    type: FlowOperationType.ADD_ACTION,
-                    request: {
-                        parentStep: step.name,
-                        stepLocationRelativeToParent: StepLocationRelativeToParent.INSIDE_TRUE_BRANCH,
-                        action: keepBaseAction(step.onSuccessAction),
-                    },
-                })
-                steps.push(...getImportOperations(step.onSuccessAction))
-            }
-        }
-        if (step.type === ActionType.LOOP_ON_ITEMS && step.firstLoopAction) {
-            steps.push({
-                type: FlowOperationType.ADD_ACTION,
-                request: {
-                    parentStep: step.name,
-                    stepLocationRelativeToParent: StepLocationRelativeToParent.INSIDE_LOOP,
-                    action: keepBaseAction(step.firstLoopAction),
-                },
-
-            })
-            steps.push(...getImportOperations(step.firstLoopAction))
-        }
-        step = step.nextAction
-    }
-    return steps
-}
-
-// It's better to use switch case, to enforce that all actions are covered
-// TODO this can be simplified
-function keepBaseAction(action: Action): Action {
-    const commonProps = {
-        name: action.name,
-        displayName: action.displayName,
-        valid: action.valid,
-    }
-    switch(action.type) {
-        case ActionType.BRANCH:
-            // PICK type and settings from action
-            return {
-                type: ActionType.BRANCH,
-                settings: action.settings,
-                ...commonProps,
-            }
-        case ActionType.LOOP_ON_ITEMS:
-            return {
-                type: ActionType.LOOP_ON_ITEMS,
-                settings: action.settings,
-                ...commonProps,
-            }
-        case ActionType.CODE:
-            return {
-                type: action.type,
-                settings: {
-                    ...action.settings,
-                    artifactPackagedId: undefined,
-                    artifactSourceId: undefined,
-                },
-                ...commonProps,
-            }
-        case ActionType.PIECE:
-            return {
-                type: action.type,
-                settings: action.settings,
-                ...commonProps,
-            }
-        case ActionType.MISSING:
-            return {
-                type: action.type,
-                settings: action.settings,
-                ...commonProps,
-            }
-    }
 }
 
 async function applySingleOperation(projectId: ProjectId, flowVersion: FlowVersion, operation: FlowOperationRequest): Promise<FlowVersion> {
@@ -561,8 +458,4 @@ async function uploadArtifact(projectId: ProjectId, codeSettings: CodeActionSett
         codeSettings.artifactPackagedId = undefined
     }
     return codeSettings
-}
-
-export const exportedFlowVersionTesting = {
-    getImportOperations,
 }
