+++ packages/shared/src/lib/flows/flow-helper.ts
@@ -90,9 +90,9 @@ function deleteAction(
 
 function getUsedPieces(trigger: Trigger): string[] {
   return traverseInternal(trigger)
-  .filter((step) => step.type === ActionType.PIECE || step.type === TriggerType.PIECE)
-  .map((step) => step.settings.pieceName)
-  .filter((value, index, self) => self.indexOf(value) === index);
+    .filter((step) => step.type === ActionType.PIECE || step.type === TriggerType.PIECE)
+    .map((step) => step.settings.pieceName)
+    .filter((value, index, self) => self.indexOf(value) === index);
 }
 
 function traverseInternal(step: Trigger | Action | undefined): (Action | Trigger)[] {
@@ -111,10 +111,10 @@ function traverseInternal(step: Trigger | Action | undefined): (Action | Trigger
   return steps;
 }
 
-
 function getAllSteps(trigger: Trigger): (Action | Trigger)[] {
   return traverseInternal(trigger);
 }
+
 function getAllParentSteps(step: Trigger): (Action | Trigger)[] {
   const steps: (Action | Trigger)[] = [];
   while (step !== undefined && step !== null) {
@@ -124,12 +124,11 @@ function getAllParentSteps(step: Trigger): (Action | Trigger)[] {
   return steps;
 }
 function getAllChildSteps(action: LoopOnItemsAction | BranchAction): (Action)[] {
-  switch(action.type)
-  {
+  switch (action.type) {
     case ActionType.LOOP_ON_ITEMS:
-    return traverseInternal(action.firstLoopAction) as Action[];
+      return traverseInternal(action.firstLoopAction) as Action[];
     default:
-      return [...traverseInternal(action.onSuccessAction),...traverseInternal(action.onFailureAction)] as Action[];
+      return [...traverseInternal(action.onSuccessAction), ...traverseInternal(action.onFailureAction)] as Action[];
   }
 
 }
@@ -195,15 +194,15 @@ function updateAction(
   }
 }
 
-function extractActions(step: Trigger | Action): { nextAction?: Action, onSuccessAction?: Action, onFailureAction?: Action , firstLoopAction?:Action } {
+function extractActions(step: Trigger | Action): { nextAction?: Action, onSuccessAction?: Action, onFailureAction?: Action, firstLoopAction?: Action } {
   const nextAction = step.nextAction;
   const onSuccessAction = step.type === ActionType.BRANCH ? step.onSuccessAction : undefined;
   const onFailureAction = step.type === ActionType.BRANCH ? step.onFailureAction : undefined;
   const firstLoopAction = step.type === ActionType.LOOP_ON_ITEMS ? step.firstLoopAction : undefined;
-  return { nextAction, onSuccessAction, onFailureAction ,firstLoopAction};
+  return { nextAction, onSuccessAction, onFailureAction, firstLoopAction };
 }
 
-function moveAction(flowVersion: FlowVersion, request: MoveActionRequest): void {
+function moveAction(flowVersion: FlowVersion, request: MoveActionRequest): FlowVersion {
   const steps = getAllSteps(flowVersion.trigger);
   const sourceStep = steps.find(step => step.name === request.name);
   if (!sourceStep || !isAction(sourceStep?.type)) {
@@ -219,12 +218,24 @@ function moveAction(flowVersion: FlowVersion, request: MoveActionRequest): void 
       params: {}
     }, `Destination step ${request.newParentStep} not found`);
   }
+  const childOperation: FlowOperationRequest[] = [];
+  const clonedSourceStep: Step = JSON.parse(JSON.stringify(sourceStep));
+  if (clonedSourceStep.type === ActionType.LOOP_ON_ITEMS || clonedSourceStep.type === ActionType.BRANCH) {
+    // Don't Clone the next action for first step only
+    clonedSourceStep.nextAction = undefined;
+    childOperation.push(...getImportOperations(clonedSourceStep));
+  }
   deleteAction(flowVersion, { name: request.name });
   addAction(flowVersion, {
     action: sourceStep as Action,
     parentStep: request.newParentStep,
     stepLocationRelativeToParent: request.stepLocationRelativeToNewParent
   });
+
+  childOperation.forEach(operation => {
+    flowVersion = flowHelper.apply(flowVersion, operation);
+  });
+  return flowVersion;
 }
 
 function addAction(flowVersion: FlowVersion, request: AddActionRequest): void {
@@ -280,7 +291,7 @@ function addAction(flowVersion: FlowVersion, request: AddActionRequest): void {
 
 function createAction(
   request: UpdateActionRequest,
-  { nextAction, onFailureAction, onSuccessAction, firstLoopAction}: { nextAction?: Action, firstLoopAction?: Action, onSuccessAction?: Action, onFailureAction?: Action},
+  { nextAction, onFailureAction, onSuccessAction, firstLoopAction }: { nextAction?: Action, firstLoopAction?: Action, onSuccessAction?: Action, onFailureAction?: Action },
 ): Action {
   const baseProperties = {
     displayName: request.displayName,
@@ -332,16 +343,16 @@ function createAction(
   return action;
 }
 
-function isChildOf(parent:LoopOnItemsAction | BranchAction,child:Action)
-{
-  switch(parent.type)
-  {
-    case ActionType.LOOP_ON_ITEMS:{
+function isChildOf(parent: LoopOnItemsAction | BranchAction, child: Action) {
+  switch (parent.type) {
+    case ActionType.LOOP_ON_ITEMS: {
       const children = getAllChildSteps(parent);
-      return children.findIndex(c=>c.name === child.name) >-1;}
-    default:{
-      const children = [...getAllChildSteps(parent),...getAllChildSteps(parent)];
-      return children.findIndex(c=>c.name === child.name) >-1;}
+      return children.findIndex(c => c.name === child.name) > -1;
+    }
+    default: {
+      const children = [...getAllChildSteps(parent), ...getAllChildSteps(parent)];
+      return children.findIndex(c => c.name === child.name) > -1;
+    }
   }
 
 
@@ -387,16 +398,117 @@ function createTrigger(
   return trigger;
 }
 
+
+export function getImportOperations(step: Action | Trigger | undefined): (FlowOperationRequest)[] {
+  const steps: FlowOperationRequest[] = []
+  while (step) {
+    if (step.nextAction) {
+      steps.push({
+        type: FlowOperationType.ADD_ACTION,
+        request: {
+          parentStep: step.name,
+          action: keepBaseAction(step.nextAction),
+        },
+      })
+    }
+    if (step.type === ActionType.BRANCH) {
+      if (step.onFailureAction) {
+        steps.push({
+          type: FlowOperationType.ADD_ACTION,
+          request: {
+            parentStep: step.name,
+            stepLocationRelativeToParent: StepLocationRelativeToParent.INSIDE_FALSE_BRANCH,
+            action: keepBaseAction(step.onFailureAction),
+          },
+        })
+        steps.push(...getImportOperations(step.onFailureAction))
+      }
+      if (step.onSuccessAction) {
+        steps.push({
+          type: FlowOperationType.ADD_ACTION,
+          request: {
+            parentStep: step.name,
+            stepLocationRelativeToParent: StepLocationRelativeToParent.INSIDE_TRUE_BRANCH,
+            action: keepBaseAction(step.onSuccessAction),
+          },
+        })
+        steps.push(...getImportOperations(step.onSuccessAction))
+      }
+    }
+    if (step.type === ActionType.LOOP_ON_ITEMS && step.firstLoopAction) {
+      steps.push({
+        type: FlowOperationType.ADD_ACTION,
+        request: {
+          parentStep: step.name,
+          stepLocationRelativeToParent: StepLocationRelativeToParent.INSIDE_LOOP,
+          action: keepBaseAction(step.firstLoopAction),
+        },
+
+      })
+      steps.push(...getImportOperations(step.firstLoopAction))
+    }
+    step = step.nextAction
+  }
+  return steps
+}
+
+// It's better to use switch case, to enforce that all actions are covered
+// TODO this can be simplified
+function keepBaseAction(action: Action): Action {
+  const commonProps = {
+    name: action.name,
+    displayName: action.displayName,
+    valid: action.valid,
+  }
+  switch (action.type) {
+    case ActionType.BRANCH:
+      // PICK type and settings from action
+      return {
+        type: ActionType.BRANCH,
+        settings: action.settings,
+        ...commonProps,
+      }
+    case ActionType.LOOP_ON_ITEMS:
+      return {
+        type: ActionType.LOOP_ON_ITEMS,
+        settings: action.settings,
+        ...commonProps,
+      }
+    case ActionType.CODE:
+      return {
+        type: action.type,
+        settings: {
+          ...action.settings,
+          artifactPackagedId: undefined,
+          artifactSourceId: undefined,
+        },
+        ...commonProps,
+      }
+    case ActionType.PIECE:
+      return {
+        type: action.type,
+        settings: action.settings,
+        ...commonProps,
+      }
+    case ActionType.MISSING:
+      return {
+        type: action.type,
+        settings: action.settings,
+        ...commonProps,
+      }
+  }
+}
+
 export const flowHelper = {
   isValid: isValid,
   apply(
     flowVersion: FlowVersion,
     operation: FlowOperationRequest
   ): FlowVersion {
-    const clonedVersion: FlowVersion = JSON.parse(JSON.stringify(flowVersion));
+    let clonedVersion: FlowVersion = JSON.parse(JSON.stringify(flowVersion));
     switch (operation.type) {
       case FlowOperationType.MOVE_ACTION:
-        moveAction(clonedVersion, operation.request);
+        clonedVersion = moveAction(clonedVersion, operation.request);
         break;
       case FlowOperationType.LOCK_FLOW:
         clonedVersion.state = FlowVersionState.LOCKED;
@@ -428,9 +540,11 @@ export const flowHelper = {
   isAction,
   getAllSteps,
   getUsedPieces,
+  getImportOperations,
   getAllSubFlowSteps,
   getStepFromSubFlow,
   isChildOf,
   getAllChildSteps,
   getAllParentSteps
 };
+
